+++ CHANGELOG.md
@@ -1,1 +1,9 @@
-## 
+## 0.1.0 - Initial Release
+
+- Initial bare bones release of this module
+- `Combat Music Management` (basic)
+  - When combat is created it does the following
+    - Opens a dialog with a list of all Combat playlists (ones that use the prefix set in settings)
+      - If you **Left Click** a playlist you can set it as the active playing playlist
+      - If you **Right Click** a playlist you can change the icon
+    - IF a Preparations playlist is set, it Will play it when a combat is created, until you select a separate playlist